requirements.txt
streamlit
pymupdf
faiss-cpu
sentence-transformers
google-genai